Adds Validation Rule format for use with Orbit
*Background*
After adding this package to an existing Laravel project by following the description in the Readme file, I started to test out the Laravel project functionality and ended up seeing errors when trying to save Models.

After some investigation, I found out that the custom Form Requests I have created to validate my form submissions were using rules that require verification of existence or uniqueness of an entry into a database table. That was done using the builtin Laravel Validation Rules `exists` and `unique`.

*Solution*
Realising that this package heavily relies in the Model itself to act, I've decided to change the rule format to the alternative Eloquent (Model) format, which uses the Eloquent fully qualified name (FQN) instead of the database table name. Doing so, we force Laravel to directly use the Model which uses Orbit behind the scenes instead of going to the database and search for the table name and make the checks.

With this change, this specific behaviour is now visible for others that may face the same situation.

*Notes*
Please, feel free to amend the text I've added to the Readme and test the solution yourself before accepting this PR.